Update app.js
1.Event Listener on Wrong Element:
The event listener is added to the wrong element. The 'submit' event should be added to the form element, not the submit button. Replace the following line:

2. String Template Literal Issue:
In the error messages for the first name and last name validation functions, you are using single quotes (') instead of backticks (`) to create a template literal. Replace these lines:

3. Form Submission: